+++ tests/fixbackend/cloud_accounts/service_test.py
@@ -30,7 +30,7 @@ from fixbackend.auth.models import User
 from fixbackend.cloud_accounts.account_setup import AssumeRoleResult, AssumeRoleResults, AwsAccountSetupHelper
 from fixbackend.cloud_accounts.models import AwsCloudAccess, CloudAccount, CloudAccountStates, GcpCloudAccess
 from fixbackend.cloud_accounts.repository import CloudAccountRepository
-from fixbackend.cloud_accounts.service_impl import CloudAccountServiceImpl
+from fixbackend.cloud_accounts.service import CloudAccountService
 from fixbackend.config import Config, ProductTierSettings
 from fixbackend.dispatcher.next_run_repository import NextRunRepository
 from fixbackend.domain_events.events import (
@@ -165,8 +165,8 @@ def service(
     http_client: AsyncClient,
     notification_service: InMemoryNotificationService,
     analytics_event_sender: AnalyticsEventSender,
-) -> CloudAccountServiceImpl:
-    return CloudAccountServiceImpl(
+) -> CloudAccountService:
+    return CloudAccountService(
         workspace_repository=workspace_repository,
         cloud_account_repository=cloud_account_repository,
         next_run_repository=next_run_repository,
@@ -188,7 +188,7 @@ def service(
 async def test_create_aws_account(
     cloud_account_repository: CloudAccountRepository,
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     user: User,
     workspace: Workspace,
     workspace_repository: WorkspaceRepository,
@@ -338,7 +338,7 @@ async def test_create_aws_account(
 
 @pytest.mark.asyncio
 async def test_delete_aws_account(
-    cloud_account_repository: CloudAccountRepository, service: CloudAccountServiceImpl, user: User, workspace: Workspace
+    cloud_account_repository: CloudAccountRepository, service: CloudAccountService, user: User, workspace: Workspace
 ) -> None:
     account = await service.create_aws_account(
         workspace_id=workspace.id,
@@ -375,7 +375,7 @@ async def test_delete_aws_account(
 
 @pytest.mark.asyncio
 async def test_store_last_run_info(
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     notification_service: InMemoryNotificationService,
     workspace: Workspace,
     next_run_repository: NextRunRepository,
@@ -420,7 +420,7 @@ async def test_store_last_run_info(
 
 @pytest.mark.asyncio
 async def test_store_last_run_info_on_error(
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     notification_service: InMemoryNotificationService,
     workspace: Workspace,
     next_run_repository: NextRunRepository,
@@ -466,7 +466,7 @@ async def test_store_last_run_info_on_error(
 
 @pytest.mark.asyncio
 async def test_get_cloud_account(
-    cloud_account_repository: CloudAccountRepository, service: CloudAccountServiceImpl, workspace: Workspace
+    cloud_account_repository: CloudAccountRepository, service: CloudAccountService, workspace: Workspace
 ) -> None:
     account = await service.create_aws_account(
         workspace_id=workspace.id,
@@ -501,7 +501,7 @@ async def test_get_cloud_account(
 
 @pytest.mark.asyncio
 async def test_list_cloud_accounts(
-    cloud_account_repository: CloudAccountRepository, service: CloudAccountServiceImpl, workspace: Workspace
+    cloud_account_repository: CloudAccountRepository, service: CloudAccountService, workspace: Workspace
 ) -> None:
     account = await service.create_aws_account(
         workspace_id=workspace.id,
@@ -527,7 +527,7 @@ async def test_list_cloud_accounts(
 
 @pytest.mark.asyncio
 async def test_update_cloud_account_name(
-    cloud_account_repository: CloudAccountRepository, service: CloudAccountServiceImpl, workspace: Workspace
+    cloud_account_repository: CloudAccountRepository, service: CloudAccountService, workspace: Workspace
 ) -> None:
     account = await service.create_aws_account(
         workspace_id=workspace.id,
@@ -568,7 +568,7 @@ async def test_update_cloud_account_name(
 async def test_handle_account_discovered_success(
     cloud_account_repository: CloudAccountRepository,
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     pubsub_publisher: RedisPubSubPublisherMock,
     workspace: Workspace,
 ) -> None:
@@ -607,7 +607,7 @@ async def test_handle_account_discovered_success(
 @pytest.mark.asyncio
 async def test_handle_account_discovered_assume_role_success(
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     account_setup_helper: AwsAccountSetupHelperMock,
     workspace: Workspace,
 ) -> None:
@@ -630,7 +630,7 @@ async def test_handle_account_discovered_assume_role_success(
 async def test_handle_account_discovered_assume_role_failure(
     cloud_account_repository: CloudAccountRepository,
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     account_setup_helper: AwsAccountSetupHelperMock,
     workspace: Workspace,
 ) -> None:
@@ -683,7 +683,7 @@ async def test_handle_account_discovered_assume_role_failure(
 async def test_handle_account_discovered_list_accounts_success(
     cloud_account_repository: CloudAccountRepository,
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     account_setup_helper: AwsAccountSetupHelperMock,
     workspace: Workspace,
 ) -> None:
@@ -726,7 +726,7 @@ async def test_handle_account_discovered_list_accounts_success(
 async def test_handle_account_discovered_list_aliases_success(
     cloud_account_repository: CloudAccountRepository,
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     account_setup_helper: AwsAccountSetupHelperMock,
     workspace: Workspace,
 ) -> None:
@@ -770,7 +770,7 @@ async def test_handle_account_discovered_list_aliases_success(
 @pytest.mark.asyncio
 async def test_enable_disable_cloud_account(
     cloud_account_repository: CloudAccountRepository,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     workspace: Workspace,
     workspace_repository: WorkspaceRepository,
     aws_marketplace_subscription: AwsMarketplaceSubscription,
@@ -870,7 +870,7 @@ async def test_enable_disable_cloud_account(
 async def test_configure_account(
     cloud_account_repository: CloudAccountRepository,
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     account_setup_helper: AwsAccountSetupHelperMock,
     workspace: Workspace,
 ) -> None:
@@ -962,7 +962,7 @@ async def test_configure_account(
 @pytest.mark.asyncio
 async def test_handle_cf_sqs_message(
     cloud_account_repository: CloudAccountRepository,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     request_handler_mock: RequestHandlerMock,
     workspace: Workspace,
 ) -> None:
@@ -1025,7 +1025,7 @@ async def test_handle_cf_sqs_message(
 
 @pytest.mark.asyncio
 async def test_move_to_degraded(
-    domain_sender: DomainEventSenderMock, service: CloudAccountServiceImpl, workspace: Workspace
+    domain_sender: DomainEventSenderMock, service: CloudAccountService, workspace: Workspace
 ) -> None:
 
     account = await service.create_aws_account(
@@ -1071,7 +1071,7 @@ async def test_move_to_degraded(
 
 @pytest.mark.asyncio
 async def test_handle_events(
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     workspace: Workspace,
     user: User,
     workspace_repository: WorkspaceRepository,
@@ -1141,7 +1141,7 @@ async def test_handle_events(
 async def test_create_gcp_account(
     cloud_account_repository: CloudAccountRepository,
     domain_sender: DomainEventSenderMock,
-    service: CloudAccountServiceImpl,
+    service: CloudAccountService,
     workspace: Workspace,
     workspace_repository: WorkspaceRepository,
     aws_marketplace_subscription: AwsMarketplaceSubscription,
